+++ src/main/java/insane96mcp/enhancedai/modules/mobs/movement/Movement.java
@@ -1,5 +1,6 @@
 package insane96mcp.enhancedai.modules.mobs.movement;
 
+import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
@@ -7,6 +8,10 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import insane96mcp.insanelib.util.MCUtils;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraftforge.common.ForgeMod;
@@ -15,22 +20,24 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.UUID;
 
-@Label(name = "Movement", description = "Makes mobs be able to move in more ways, like climbing or swim.")
+@Label(name = "Movement", description = "Makes mobs be able to move in more ways, like climbing or swim. Only mobs in the enhancedai:allow_climbing entity type tag are allowed to climb.")
 @LoadFeature(module = Modules.Ids.MOBS)
 public class Movement extends Feature {
+    public static final TagKey<EntityType<?>> ALLOW_CLIMBING = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "allow_climbing"));
+    //public static final TagKey<EntityType<?>> ALLOW_TARGETING_LADDERS = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "allow_climbing"));
 
     final UUID UUID_SWIM_SPEED_MULTIPLIER = UUID.fromString("6d2cb27e-e5e3-41b9-8108-f74131a90cce");
 
     @Config
     @Label(name = "Allow Climbing", description = "If true, mobs will be able to climb (up and down)")
     public static Boolean allowClimbing = true;
-    @Config
+    /*@Config
     @Label(name = "Target Ladders", description = "If true, mobs try to find climbable blocks to reach the target")
-    public static Boolean targetLadders = false;
+    public static Boolean targetLadders = false;*/
 
     @Config(min = 0d, max = 4d)
-    @Label(name = "Swim Speed Multiplier", description = "How faster mobs can swim. Setting to 0 will leave the swim speed as vanilla. I recommend using mods like Mobs Properties Randomness to have more control over the attribute.")
-    public static Double swimSpeedMultiplier = 2.5d;
+    @Label(name = "Swim Speed Addition Multiplier", description = "How faster mobs can swim. Setting to 0 will leave the swim speed as vanilla. I recommend using mods like Mobs Properties Randomness to have more control over the attribute.")
+    public static Double swimSpeedAdditionMultiplier = 2.5d;
 
     public Movement(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
@@ -46,12 +53,12 @@ public class Movement extends Feature {
             mob.goalSelector.addGoal(3, new ClimbLaddersGoal(mob));
         }
 
-        if (targetLadders) {
+        /*if (targetLadders) {
             mob.targetSelector.addGoal(0, new FindLaddersGoal(mob));
-        }
+        }*/
 
-        if (swimSpeedMultiplier != 0d) {
-            MCUtils.applyModifier(mob, ForgeMod.SWIM_SPEED.get(), UUID_SWIM_SPEED_MULTIPLIER, "Enhanced AI Swim Speed Multiplier", swimSpeedMultiplier, AttributeModifier.Operation.MULTIPLY_BASE, false);
+        if (swimSpeedAdditionMultiplier != 0d) {
+            MCUtils.applyModifier(mob, ForgeMod.SWIM_SPEED.get(), UUID_SWIM_SPEED_MULTIPLIER, "Enhanced AI Swim Speed Multiplier", swimSpeedAdditionMultiplier, AttributeModifier.Operation.MULTIPLY_BASE, false);
         }
     }
 }