+++ src/main/java/insane96mcp/enhancedai/modules/shulker/ShulkerArmor.java
@@ -1,19 +1,26 @@
 package insane96mcp.enhancedai.modules.shulker;
 
+import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.monster.Shulker;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
 
 import java.util.UUID;
 
-@Label(name = "Shulker Armor")
+@Label(name = "Shulker Armor", description = "Use the enhancedai:apply_armor_modifiers to add more shulkers that are affected by this feature.")
 @LoadFeature(module = Modules.Ids.SHULKER)
 public class ShulkerArmor extends Feature {
+    public static final TagKey<EntityType<?>> APPLY_ARMOR_MODIFIERS = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "apply_armor_modifiers"));
     @Config(min = 1, max = 40)
     @Label(name = "Armor when closed")
     public static Double armorWhenClosed = 30d;
@@ -39,5 +46,9 @@ public class ShulkerArmor extends Feature {
         CLOSED_MODIFIER = new AttributeModifier(COVERED_ARMOR_MODIFIER_UUID, "Covered armor bonus", armorWhenClosed, AttributeModifier.Operation.ADDITION);
         PEEK_MODIFIER = new AttributeModifier(COVERED_ARMOR_MODIFIER_UUID, "Covered armor bonus", armorWhenPeeking, AttributeModifier.Operation.ADDITION);
         OPEN_MODIFIER = new AttributeModifier(COVERED_ARMOR_MODIFIER_UUID, "Covered armor bonus", armorWhenOpen, AttributeModifier.Operation.ADDITION);
+    }
+
+    public static boolean isAffectedByArmorModifers(Shulker shulker) {
+        return Feature.isEnabled(ShulkerArmor.class) && shulker.getType().is(APPLY_ARMOR_MODIFIERS);
     }
 }